refactor: split out example code and deps
